# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title:  DevOps Ecosystem Platform — Combined API
  version: "2.0.0"
  description: |
    API contract for 4 modules added in BRD-001 v1.1 (FR-019 – FR-032).
    Auth: Bearer JWT (ADR-003). All write operations produce audit log (NFR-007).
    Banking-grade: all endpoints except /health require Authorization header.
  contact:
    name: Backend Dev (Agent 06)
servers:
  - url: http://localhost:8001
    description: PPG System — Annual Plans, Project Objects, Reports
  - url: http://localhost:8002
    description: BA Workflow — BA Documents v2
  - url: http://localhost:8003
    description: Test Platform — Test Documents
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/documents:
    get:
      summary: List BA documents (FR-027, FR-028, FR-029)
      operationId: listBADocuments
      tags: [ba-documents]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: doc_type
          in: query
          schema:
            type: string
            enum: [BRD, BRS, FSD, API_SPEC, ERD, DATA_DICT, WIREFRAME, PROCESS_FLOW]
        - name: status
          in: query
          schema:
            type: string
        - name: object_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by linked object (FR-027)
        - name: milestone_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of BA documents

    post:
      summary: Create BA document (FR-027, FR-029)
      operationId: createBADocument
      tags: [ba-documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, doc_type, title]
              properties:
                project_id:
                  type: string
                  format: uuid
                doc_type:
                  type: string
                  enum: [BRD, BRS, FSD, API_SPEC, ERD, DATA_DICT, WIREFRAME, PROCESS_FLOW]
                title:
                  type: string
                  maxLength: 300
                content:
                  type: string
                  default: ""
                milestone_id:
                  type: string
                  format: uuid
                  nullable: true
                object_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                metadata:
                  type: object
      responses:
        "201":
          description: Created document
        "422":
          description: OBJECT_INCOMPLETE or validation error (BR-011)
...